Gbe 63: upgrade python, cleanup tests (#116)
* got to manage.py startup

* python 3.7 and init updates

* first page load - landing & some scheduling

* Tests actually run

* Screw it, 2to3 everything in sight

* get tests to pass - first 4 + 1 bug fix

* volunteer wizard passes.  *.bak is now annoying

* 2 tests, 2 pepdiff

* pepdiff and another test passing

* reporting/staff_area + all email + more pepdiff

* 4 more scheduling tests

* pepdiff mostly done, all of scheduling

* more tests

* more test cleanup, 1 bug (edit_troupe)

* done?

* 2 bugs in export calendar, removed test from edit_volunteer

* configure baseproject default_v3 4.7

* coverage + deploy error

* fixing diff issue

* configure baseproject default_v3_py3 4.16

+++ gbe/scheduling/forms/schedule_occurrence_form.py
@@ -25,8 +25,8 @@ from settings import (
 
 time_start = 8 * 60
 time_stop = 24 * 60
-conference_times = [(time(mins/60, mins % 60),
-                     time(mins/60, mins % 60).strftime(GBE_TIME_FORMAT))
+conference_times = [(time(int(mins/60), mins % 60),
+                     time(int(mins/60), mins % 60).strftime(GBE_TIME_FORMAT))
                     for mins in range(time_start, time_stop, 30)]
